updated Figure SM4 averaged AE
